dx references: fix wrong object problem.
The problem is that when source and target Plone sites run on the same
Zope instance dexterity relations may be looked up wrong.
When an the relation target is not yet on the target site it is
found from the target site on the source site when queried with the
absolute path.
Therefore we should always use relative paths when putting paths on the wire.