ftw.simplelayout: remove deleted contentish children.
The content of simplelayout pages consist of blocks which have usually
no workflow and share the publishing cycle with the page.

In order to share the publishing cycle correctly, we need to not publish
any delete jobs for (block-) objects which are considered simplelayout
content.

For solving this problem, a `is_sl_contentish` function is introduced,
which decides which objects are part of the simplelayout content of a
simplelayout container.
This checks may be recursive since there are folderish simplelayout
blocks, such as the listingblock, which may contain non-simplelayout
objects.
When those nested objects do not have a workflow, they share the
publishing cycle of the simplelayout container (parent of the parent)
and are thus considered simplelayout contentish.

The ``RemoveDeletedSLContentishChildren`` is used for both, simplelayout
containers as well as (folderish) simplelayout blocks and extracts the
UUIDs of all contentish children.
When receiving an object the UUIDs are then compared with the state on
the receiver and blocks which no longer exist on the sender side are
deleted from the receiver side.