Fix daylight savings issue in p.a.events' start / end.
Because plone.app.event's IEventBasic has a separate timezone field,
it does some strange things with FakeZone on start / end.
This leads to problems when publishing with the regulare publisher
field adapter.

In order to fix to those issues we handle start / end dates in this
adapter explicitly, so that we have full control.