Improve publishing the layout property on plone site root.
Publishing the layout property on plone site root has been introduced in 2.7.3 but used to fail if the plone site root on the receiving end does not have the layout property.

We must therefor check if the property is available before trying to delete it in order to prevent the following error raised in `manage_delProperties()`: BadRequest: The property <em>layout</em> does not exist.