Drop support for ftw.contentpage.
ftw.contentpage is no longer maintainend and replaced with
ftw.simplelayout in most projects.

Since ftw.contentpage is no longer developed, it is not compatible
with the newest ftw.testing 2.0.0 release.

In order to keep ftw.publisher up to date we need to drop support for
ftw.contentpage so that we can update ftw.testing.