Disable plone.app.event fix for Plone 5.
In Plone 4, plone.app.event used to separate the timezone info from
the dates, which made it necessary to have a custom publisher adapter
for fixing the start and end values.

In Plone 5, plone.app.event has a different implementation which
breaks when the fix is used because the fix uses Zope DateTime which
is no longer wise.