Fix timezone transfer from sender to receiver
Two problems existed:
- The datetime was not always transferred as ISO-8601 formatted string
- The timezone was converted to GTM on the receiving side. The `pytz`
  module does not understand GMT timezones except for GMT+0 (which is why
  it worked until now)

Possible solutions:
- Correct the timezone to something understandable by `pytz` on the
  receiver side.
- Send the timezone separately as something that `pytz` already
  understands.

We choose solution two because it is easier to implement. For solution 1
we would have had to create our own `GMT` parser that makes it `pytz`
compatible or get another package to do it.

With the second solution we just strip the timezone and send separately
to get a string on the receiver side which `pytz` already understands.
This however means we need to change both sides and add additional
conditions for `Date` and `Time` objects.

We stripped the datetime object's timezone and did not recalculate it to
UTC because it would require additional conditions. That is because not
all timezone aware objects implement a recalculating method like the
`datetime.datetime` objects `.astimezone(...)`.

In addition to the timezone we now also make sure that when we get a
date (not a datetime) that it is set as date again. The same for time
objects. Before these changes the receiver created used a datetime
object for those values too instead of a simple date and a time object.